Refactor check_for_mut_range_bound to its own module
